Network interface: updated kotlin typing of precalculateLimitUpdate

+++ socket-interface/src/main/kotlin/com/github/prologdb/net/session/ProtocolMessage.kt
@@ -54,7 +54,7 @@ data class ConsumeQuerySolutionsCommand(
     val queryId: Int,
     val amount: Int?,
     val closeAfterwards: Boolean,
-    val precalculateLimitUpdate: Int?,
+    val precalculateLimitUpdate: Long?,
     val notifyAboutClose: Boolean,
     val handling: SolutionHandling
 ) : ProtocolMessage() {
